Moved Guides´ routing to a seperate module By moving routing to a seperate module, routing to default route in giudes works
