+++ apps/cookbook/src/app/guides/guides.module.ts
@@ -11,6 +11,7 @@ import { CodeViewerModule } from '../shared/code-viewer/code-viewer.module';
 import { GridLayoutExtendedComponent } from './grid-layout/grid-layout-extended/grid-layout-extended.component';
 import { GridLayoutMultipleContainersComponent } from './grid-layout/grid-layout-multiple-containers/grid-layout-multiple-containers.component';
 import { GridLayoutSingleContainerComponent } from './grid-layout/grid-layout-single-container/grid-layout-single-container.component';
+import { GuidesRouterModule } from './guides-routing.module';
 import { GuidesComponent } from './guides.component';
 import { VirtualScrollListComponent } from './virtual-scroll/virtual-scroll-list/virtual-scroll-list.component';
 
@@ -22,33 +23,11 @@ const guidesComponents = [
   VirtualScrollListComponent,
 ];
 
-const routes = [
-  {
-    path: '',
-    component: GuidesComponent,
-  },
-  {
-    path: 'virtual-scroll-list',
-    component: VirtualScrollListComponent,
-  },
-  {
-    path: 'grid-layout-single-container',
-    component: GridLayoutSingleContainerComponent,
-  },
-  {
-    path: 'grid-layout-multiple-containers',
-    component: GridLayoutMultipleContainersComponent,
-  },
-  {
-    path: 'grid-layout-extended',
-    component: GridLayoutExtendedComponent,
-  },
-];
-
 @NgModule({
   imports: [
     CommonModule,
-    RouterModule.forChild(routes),
+    GuidesRouterModule,
+    RouterModule,
     KirbyModule,
     IphoneModule,
     CodeViewerModule,
